+++ Connectors/StdDevConnector.cpp
@@ -1,25 +1,27 @@
 #include "StdDevConnector.hpp"
 #include <algorithm>
 #include <math.h>
-void	StdDevConnector::setValue(int64_t AnalysisValue)
+#include <iostream>
+void	StdDevConnector::setValue(int16_t AnalysisValue)
 {
     frames[current] = AnalysisValue;
 
     current = (current + 1) % maxFrames;
     auto avg = 0;
     auto sse = 0;
-    std::for_each(frames.begin(), frames.end(), [&] (int64_t n) {
+    std::for_each(frames.begin(), frames.end(), [&] (int16_t n) {
         avg += (n / maxFrames);
     });
-    std::for_each(frames.begin(), frames.end(), [&] (int64_t n) {
+    std::for_each(frames.begin(), frames.end(), [&] (int16_t n) {
         sse += (n - avg)*( n - avg);
     });
 
     value = sqrt(sse/(maxFrames-1));
+    //std::cout << "value " << value << std::endl;
 
 }
 
-StdDevConnector::StdDevConnector(int64_t frameMax){
+StdDevConnector::StdDevConnector(int16_t frameMax){
     current = 0;
     maxFrames = frameMax;
     for( int i = 0; i < frameMax; i++){
@@ -27,6 +29,6 @@ StdDevConnector::StdDevConnector(int64_t frameMax){
     }
 }
 
-int32_t StdDevConnector::getValue(){
+int16_t StdDevConnector::getValue(){
     return value;
 }
