+++ include/StdDevConnector.hpp
@@ -3,9 +3,9 @@
 
 class StdDevConnector : public Connector {
     public:
-        StdDevConnector(int64_t n);
-        void setValue(int64_t AnalysisValue);
-        int32_t getValue();
+        StdDevConnector(int16_t n);
+        void setValue(int16_t AnalysisValue);
+        int16_t getValue();
     protected:
         std::vector<int> frames;
         int current;
